requirements.txt
absl-py==1.4.0
cachetools==5.3.0
ccimport==0.4.2
certifi==2023.5.7
charset-normalizer==3.1.0
cloudpickle==2.2.1
cumm-cu117==0.4.9
cycler==0.11.0
entrypoints==0.4
Farama-Notifications==0.0.4
fire==0.5.0
fonttools==4.38.0
google-auth==2.19.0
google-auth-oauthlib==0.4.6
grpcio==1.54.2
gymnasium==0.28.1
h5py==3.8.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
jax-jumpy==1.0.0
kiwisolver==1.4.4
lark==1.1.5
llvmlite==0.39.1
Markdown==3.4.3
MarkupSafe==2.1.2
mathutils==2.81.2
matplotlib==3.5.3
ninja==1.11.1
numba==0.56.4
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
packaging==23.1
pccm==0.4.7
Pillow==9.5.0
portalocker==2.7.0
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.10.4
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.7.3
six==1.16.0
spconv-cu117==2.3.6
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.3.0
tianshou==0.5.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
typing_extensions==4.6.2
urllib3==1.26.16
Werkzeug==2.2.3
zipp==3.15.0
wandb
